chore: memo the future plan
